+++ app/Borrowing.php
@@ -6,5 +6,14 @@ use Illuminate\Database\Eloquent\Model;
 
 class Borrowing extends Model
 {
-    //
+    protected $fillable = [
+        'inventory_item_id',
+        'borrower_id',
+        'initial_lender_id',
+        'guarantee',
+        'status_id',
+        'start_date',
+        'expected_return_date',
+        'notes_before'
+    ];
 }