+++ database/migrations/16_09_2018_000002_create_users_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
+use App\UserRole;
 
 class CreateUsersTable extends Migration
 {
@@ -21,8 +22,8 @@ class CreateUsersTable extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->boolean('admin')->default(false);
-            $table->boolean('lender')->default(false);
+            $table->unsignedInteger('role')->default(UserRole::AUCUN);
+            $table->foreign('role')->references('id')->on('user_roles');
             $table->rememberToken();
             $table->timestamps();
         });