+++ routes/web.php
@@ -6,7 +6,7 @@ Route::get('/', function () {
 Route::get('/new-borrowing', 'NewBorrowingController@index');
 Route::post('/new-borrowing', 'NewBorrowingController@store');
 Route::get('/end-borrowing', 'EndBorrowingController@index');
-Route::get('/borrowings-history', 'BorrowingsHistoryController@index');
+Route::get('/borrowings-history', 'BorrowingsHistoryController@index')->name('borrowings-history');
 Route::get('/view-inventory', 'ViewInventoryController@index');
 Route::get('/edit-inventory', 'EditInventoryController@index');
 Route::get('/admin', 'AdminController@index');