+++ changelog.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.5.0] - Unreleased
+## [0.5.0] - 2018-09-16
 
 ### Added
 
@@ -17,15 +17,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     * Checkout filling.
     * Inventory request.
     * Fuzzy games search.
+    * Borrowing post.
 * Controllers
     * Templates.
+    * New borrowing store function.
+* Requests
+    * New borrowing validation.
 * Database
     * Access configuration.
     * Tables creation.
     * Reference tables seeding.
-    * Inventory items seeding for local environment.
+    * Inventory items and users seeding for local environment.
 * Public
-    * CSS for input fields in modals.
+    * CSS for new borrowing view.
 
 ### Changed